// src/gs-api/src/model/commandeClientResponseDto.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LigneCommandeClientResponseDto } from './ligneCommandeClientResponseDto';


export interface CommandeClientResponseDto { 
    id?: number;
    code?: string;
    dateCommande?: string;
    entrepriseId?: number;
    ligneCommandeClients?: Array<LigneCommandeClientResponseDto>;
}