// src/gs-api/src/model/commandeFournisseurResponseDto.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LigneCommandeFournisseurResponseDto } from './ligneCommandeFournisseurResponseDto';
import { FournisseurResponseDto } from './fournisseurResponseDto';


export interface CommandeFournisseurResponseDto { 
    id?: number;
    code?: string;
    dateCommande?: string;
    fournisseur?: FournisseurResponseDto;
    entrepriseId?: number;
    ligneCommandeFournisseurs?: Array<LigneCommandeFournisseurResponseDto>;
}